Refactor left/right turn callbacks
I felt that using operator.add and operator.sub to traverse the
compass_points list was quite hard to follow, particularly given the
variable names that I'd assigned them to.

Passing a callable that knows how to return the new index after the turn
feels a bit clearer, primarily because it's more obvious how to assign
useful names to the things involved!